bcrypto: fix EVP bug. see bcoin-org/bcoin#628.
At some point during the transition from node 8 to node 10 a certain change was
applied to how the default crypto module is loaded. In the node 8 days,
`crypto` had to be required in order to initialize OpenSSL's EVP interface. Our
modules which utilize EVP are: pbkdf2, cipher, and rsa.

The way node.js initializes EVP is on the crypto module load, in the
`InitCryptoOnce()` call. The function invokes:

```
SSL_library_init();
OpenSSL_add_all_algorithms();
```

Which loads all ciphers and hash algorithms in order to be used by NID/name
through the EVP interface.

I was aware of this oddity before, but I seemed to have forgotten about it by
the time node 10 came out. Node 10, for whatever reason, doesn't need to need
to be initialized by calling `require('crypto')` (if anyone can figure out why,
that would be helpful).

We can solve this by lazily requiring the crypto module (if major_version < 10)
in the appropriate places.

This commit also adds a POST file before running tests and will immediately
choke if it detects some kind of inconsistency with EVP algorithm loading. On
boot, bcrypto will also warn the user if they're on a different node version
than the one they compiled against.