schnorr: implement publicKeyTweakTest()
This method is analogous to secp256k1_xonly_pubkey_tweak_test()
in the branch of libsecp256k1 used in the bip-taproot working branch:
https://github.com/sipa/bitcoin/blob/e97f84c8044b4f095e16956266da9c3416dd639d/src/secp256k1/src/secp256k1.c#L839

Unlike publicKeyTweakAdd() this method preserves the Y coordinate of
the tweaked key, and negates it based on a provided boolean argument.
The tweaked-and-possibly-negated key is then tested for equality to
a second public key.

This algorithm is defined in bip-taproot as a witness commitment check
during script-path spending.